pyproject.toml
[tool.poetry]
name = "case"
version = "0.1.0"
description = "CASE - IOT Server"
authors = ["Anderson Bairo <ander_slash@yahoo.com.br>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
djangorestframework = "^3.12.4"
python-decouple = "^3.5"
psycopg2 = "^2.9.1"
djangorestframework-simplejwt = "4.8.0"
PyJWT = "2.1.0"
paho-mqtt = "^1.5.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
ipdb = "^0.13.9"
ipython = "^7.28.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
